v1.09: fixed bug with daemon dying after trying to parse wrong request
git-svn-id: http://scm.tomtomgroup.com/munin-plugins/trunk@55 7b71e12a-a62d-4719-a91f-e5bf40180492

+++ scripts/pre-uninstall.sh
@@ -1,2 +1,5 @@
+if [ "$1" = 0  ]; then
+/sbin/service munindaemon stop
 rm -rf /var/log/munindaemon
 rm -rf /var/run/munindaemon
+fi
